telegram_bot/requirements.txt
torch==1.13.1
opencv-python==4.5.3.56
numpy==1.24.0
python-telegram-bot==20.0
Pillow==9.4.0
albumentations==1.3.0
albumentations-pytorch==0.0.10
